removed JS files to designated direcories
