+++ account.html
@@ -147,7 +147,7 @@
 			<p class="disclaimer-info__text">Email нужен для восстановления пароля</p>
 		</div>
 	</main>
-	<script src="./scripts/script.js" defer></script>
+	<script src="./scripts/account.js" defer></script>
 </body>
 
 </html>
