+++ index.html
@@ -57,7 +57,7 @@
 	</template>
 
 
-	<script src="./scripts/script.js" type="module"></script>
+	<script src="./scripts/index.js" type="module"></script>
 </body>
 
 </html>
